+++ src/parser/part/usage.rs
@@ -1403,6 +1403,10 @@ fn part_usage_body_element(input: Input<'_>) -> IResult<Input<'_>, Node<PartUsag
             let elem = PartUsageBodyElement::PortUsage(Box::new(usage));
             return Ok((next, node_from_to(start, next, elem)));
         }
+        if let Ok((next, usage)) = analysis_case_usage(start) {
+            let elem = PartUsageBodyElement::AnalysisCaseUsage(usage);
+            return Ok((next, node_from_to(start, next, elem)));
+        }
     }
     let frag = start.fragment();
     let first_30 = frag.get(..30.min(frag.len())).unwrap_or(frag);